requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
itsdangerous==0.24
python-dateutil==2.4.2
requests==2.6.0
six==1.9.0
wsgiref==0.1.2
